refactor: adjust router for authentication
